[ao] Added README for ModelReport functionality
Summary: This adds a README for the ModelReport functionality that
contains an overview of the class, what it does,
and how it works, an example of usage, information on how to implement a
new detector (since this is how core functionality is added), folder
structure information, and finally information on tests and where they
are located.

The ModelReport class is still in development and will, in the future,
get additional features such as visualizations, and the README will be
updated with this information as it is added.

Test Plan: Just a new README, no code is added, README will be reviewed
for accuracy and ease of use/ easiness to read.

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: a8830730882a009fa495347671f8b16d72117797
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81369