[ao][sparsity] Basic implementation of activation sparsifier
The Activation sparsifier class aims to sparsify/prune activations in a neural
network. The idea is to attach the sparsifier to a layer (or layers) and it
zeroes out the activations based on the mask_fn (or sparsification function)
input by the user.
The mask_fn is applied once all the inputs are aggregated and reduced i.e.
mask = mask_fn(reduce_fn(aggregate_fn(activations)))

Note::
    The sparsification mask is computed on the input **before it goes through the attached layer**.

Test Plan:
```python test/test_ao_sparsity.py TestActivationSparsifier```

ghstack-source-id: 519e0ed2c3762882431f396a2a035f6e5bba1b18
Pull Request resolved: https://github.com/pytorch/pytorch/pull/80886